Creates a visual graph of a given network

baseNet helps provide users with clear, concise visualizations of their standard network graphs so that they can perform an effective analysis of the network.

baseNet(data, nodes = NULL, layout = "lgl", label, label_color,
  directed = F, node_color, node_size, edge_color, edge_size)

Arguments

data
A network object that can be converted to a tidygraph tbl_graph (acceptable classes include 'data.frame', 'igraph', 'list', 'matrix', and 'network'). See ?tbl_graph for more information about accepted classes of networks.

nodes
A data.frame containing information about the nodes in the graph. Default: NULL

layout
The type of layout used to plot the network using ggraph. Either a valid string, a function, a matrix, or a data.frame. See ?ggraph for more information. Default: "lgl"

label
The variable from the nodes table that should be used to label the nodes in the plotted network. Default: first column of the nodes table

label_color
The color of the label. Default: "navy"

directed
Indicates if the constructed graph should be directed. Default: F

node_color
The color of the nodes of the network. Accepted inputs are either a character vector that notes a specific color (ex."red") or the name of a variable from the nodes table. Default: "cornflowerblue"

node_size
The size of the nodes of the network. Accepted inputs are either a numeric vector that notes a specific size (ex. 6) or the name of a variable from the nodes table. Default: 8

edge_color
The color of the edges of the network. Accepted inputs are either a character vector that notes a specific color (ex."red") or the name of a variable from the edges table. Default: "slategrey"

edge_size
The size/width of the edges of the network. Accepted inputs are either a numeric vector that notes a specific size (ex. 0.5) or the name of a variable from the nodes table. Default: 0.6

Value

A list, k, containing the network data in the form of a tidygraph, k$data, and a ggmap object of the plotted network, k$network.

Details

The baseNet function converts the user's network into a tidygraph class object, returning the both the tidygraph object and plotting the network using ggmap. This allows for the user to futher edit the network object as needed as well as allowing them to add further modifications to the ggmap plot, such as editing the title. In order to customize the nodes of the network using a measure of centrality, the user must input one of the values featured bellow into the node_color/node_size aruguments, which correspond with following tidygraph centrality measurement functions:

betweenness: centrality_betweenness()

closeness: centrality_closeness()

degree: centrality_degree()

eigen: centrality_eigen()

Users also have the option of customizing the edges of their network using the variable edge_weight, which notes the weight between two edges.

Examples

if (FALSE) {
library(igraphdata)
data("karate")
baseNet(karate, label=name, layout="nicely")
}