Creates a graph of a geographic network

geoNet accepts a network in the form of a set of nodes and edges, where each node corresponds to an actual geographic location. It returns an interactive geographic map with this network overlaid on it. It also allows the user to specify other variables according that modify the color and size of the nodes and edges of the network.

geoNet(df, geodata = NULL, node_color = "red", node_size = 3,
  edge_color = "navy", edge_size = 4, name = "name", lat = "lat",
  lng = "lng", orig = "origin", dest = "destination")

Arguments

df
This is the primary data set. It corresponds to the edges of the network. It accepts an igraph or tidygraph object, or data frames or an adjacency matrices. If this is an igraph or tidygraph object, the geodata parameter would be ignored as all the node attributes (including the geographic data) would be extracted from the object itself. It this is a data frame or and adjacency matrix, then geodata cannot be NULL.

geodata
This is a data frame that describes the attributes of the nodes, including its geographic attributes. Default: NULL

node_color
This accepts a character vector that is either a color or the name of the variable amongst the node attributes that should be used to decide the color of each node. The variable name should correspond to a variable that is either a factor variable or is numeric. Default: 'red'

node_size
This accepts either an integer or a character vector that is the name of the variable amongst the node attributes that should be used to decide the size of each node. The variable name should correspond to a variable that is numeric in nature. Default: 3

edge_color
This accepts a character vector that is either a color or the name of the variable amongst the edge attributes that should be used to decide the color of each edge The variable name should correspond to a variable that is either a factor variable or is numeric. Default: 'navy'

edge_size
This accepts either an integer or a character vector that is the name of the variable amongst the edge attributes that should be used to decide the width of each edge. The variable name should correspond to a variable that is numeric in nature. Default: 4

name
This accepts a character vector that is the name of the variable amongst the node attributes that describes the name of each node. Default: 'name'

lat
This accepts a character vector that is the name of the variable amongst the node attributes that describes the latitude of each node. Default: 'lat'

lng
This accepts a character vector that is the name of the variable amongst the node attributes that describes the longitude of each node. Default: 'lng'

orig
This accepts a character vector that is the name of the variable amongst the edge attributes that describes one end of each edge. Default: 'origin'

dest
This accepts a character vector that is the name of the variable amongst the edge attributes that describes the second end of each edge. Default: 'destination'

Value

This function returns an interactive map with the network on it. This map is of the class Leaflet.

Details

If users wish to customize their geoNet graph based on measures of network centrality, unless they have this information on centrality measures for their network already featured in the nodes/edges tables of the network, users should input their network into the baseNet function and then use the tidygraph "data" output as the data argument for the geoNet function, making sure to specify the modified node/edge arguments as necessary.

Examples

if (FALSE) {
data(airports)
data(Southwest)

sw_nodes = unique(c(Southwest$ORIGIN, Southwest$DEST))
k2 = airports$AIRPORT %in% sw_nodes
airports=as.data.frame(airports[k2,])
flights<-subset(Southwest, select=c(3,4,2))

m<-geoNet(df = flights, geodata = airports, node_color = "red", node_size = 3,
edge_color = 'cornflowerblue', edge_size = 4, name = "AIRPORT", lat = "LATITUDE", lng = "LONGITUDE",
orig = "ORIGIN", dest = "DEST")

m
}